requirements.txt
flask~=3.0.2
flask-wtf
flask-bootstrap
requests~=2.31.0
pillow~=10.2.0
numpy~=1.26.4
matplotlib~=3.8.3
waitress~=3.0.0
WTForms~=3.1.2
Werkzeug~=3.0.1